refactor(application): split 8 services into partial classes by concern (#279)
