+++ LgymApi.Application/TrainerRelationships/TrainerRelationshipService.Links.cs
@@ -1,0 +1,36 @@
+using LgymApi.Application.Common.Errors;
+using LgymApi.Application.Common.Results;
+using LgymApi.Domain.ValueObjects;
+using LgymApi.Resources;
+using UserEntity = LgymApi.Domain.Entities.User;
+
+namespace LgymApi.Application.Features.TrainerRelationships;
+
+public sealed partial class TrainerRelationshipService
+{
+    public async Task<Result<Unit, AppError>> UnlinkTraineeAsync(UserEntity currentTrainer, Id<UserEntity> traineeId, CancellationToken cancellationToken = default)
+    {
+        var linkResult = await EnsureTrainerOwnsTraineeAsync(currentTrainer, traineeId, cancellationToken);
+        if (linkResult.IsFailure)
+        {
+            return Result<Unit, AppError>.Failure(linkResult.Error);
+        }
+
+        await _trainerRelationshipRepository.RemoveLinkAsync(linkResult.Value, cancellationToken);
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
+        return Result<Unit, AppError>.Success(Unit.Value);
+    }
+
+    public async Task<Result<Unit, AppError>> DetachFromTrainerAsync(UserEntity currentTrainee, CancellationToken cancellationToken = default)
+    {
+        var link = await _trainerRelationshipRepository.FindActiveLinkByTraineeIdAsync(currentTrainee.Id, cancellationToken);
+        if (link == null)
+        {
+            return Result<Unit, AppError>.Failure(new TrainerRelationshipNotFoundError(Messages.DidntFind));
+        }
+
+        await _trainerRelationshipRepository.RemoveLinkAsync(link, cancellationToken);
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
+        return Result<Unit, AppError>.Success(Unit.Value);
+    }
+}
